House keeping (#16)
* update frontend dependencies. Update eslint and tsconfig

* update backend dependencies

* add ruff and mypy

* fix some ruff issues (sorting)

* fix ruff issues

* update config

* fix mypy errors

* fix mypy issues

* fix frontend lint issues

* fix security issue with fast-api-sso

* add pre-commit config file

* remove facebook sso

* remove deprecated Grid component. Use Grid2 instead

* fix tests by changing testing api url

* update node and mongodb versions in tests workflow

* make black use pyproject.toml in github actions

* Add config to black

* change working dir for python lint

* refactor docker-compose setup. Update traefik to v3.2

* add docker watch for frontend container

* add node_modules ignore to docker watch

* refactor config to use root level .env file

* update docker doc in README

+++ backend/app/routers/users.py
@@ -1,17 +1,17 @@
-from typing import List, Optional, Any
+from typing import Any, List, Optional
 from uuid import UUID
 
-from fastapi import APIRouter, HTTPException, Body, Depends
-from pymongo import errors
-from pydantic.networks import EmailStr
 from beanie.exceptions import RevisionIdWasChanged
+from fastapi import APIRouter, Body, Depends, HTTPException
+from pydantic.networks import EmailStr
+from pymongo import errors
 
+from .. import models, schemas
 from ..auth.auth import (
-    get_hashed_password,
     get_current_active_superuser,
     get_current_active_user,
+    get_hashed_password,
 )
-from .. import schemas, models
 
 router = APIRouter()
 
@@ -75,9 +75,7 @@ async def update_profile(
     )
     try:
         if update_data["password"]:
-            update_data["hashed_password"] = get_hashed_password(
-                update_data["password"]
-            )
+            update_data["hashed_password"] = get_hashed_password(update_data["password"])
             del update_data["password"]
     except KeyError:
         pass
@@ -118,12 +116,14 @@ async def update_user(
         the current superuser, by default Depends(get_current_active_superuser)
     """
     user = await models.User.find_one({"uuid": userid})
+    if user is None:
+        raise HTTPException(status_code=404, detail="User not found")
     if update.password is not None:
         update.password = get_hashed_password(update.password)
-    user = user.model_copy(update=update.model_dump(exclude_unset=True))
+    updated_user = user.model_copy(update=update.model_dump(exclude_unset=True))
     try:
-        await user.save()
-        return user
+        await updated_user.save()
+        return updated_user
     except errors.DuplicateKeyError:
         raise HTTPException(
             status_code=400, detail="User with that email already exists."
@@ -160,5 +160,7 @@ async def delete_user(
     userid: UUID, admin_user: models.User = Depends(get_current_active_superuser)
 ):
     user = await models.User.find_one({"uuid": userid})
+    if user is None:
+        raise HTTPException(status_code=404, detail="User not found")
     await user.delete()
     return user
